Move metadata builder actor classes into 'services'
