+++ services/src/main/scala/cromwell/services/metadata/impl/MetadataServiceActor.scala
@@ -109,7 +109,7 @@ final case class MetadataServiceActor(serviceConfig: Config, globalConfig: Confi
     case listen: Listen => writeActor forward listen
     case v: ValidateWorkflowIdInMetadata => validateWorkflowIdInMetadata(v.possibleWorkflowId, sender())
     case v: ValidateWorkflowIdInMetadataSummaries => validateWorkflowIdInMetadataSummaries(v.possibleWorkflowId, sender())
-    case action: ReadAction => readActor forward action
+    case action: MetadataReadAction => readActor forward action
     case RefreshSummary => summaryActor foreach { _ ! SummarizeMetadata(metadataSummaryRefreshLimit, sender()) }
     case MetadataSummarySuccess => scheduleSummary()
     case MetadataSummaryFailure(t) =>
