+++ services/src/main/scala/cromwell/services/metadata/impl/builder/MetadataBuilderActor.scala
@@ -1,4 +1,4 @@
-package cromwell.webservice.metadata
+package cromwell.services.metadata.impl.builder
 
 import java.time.OffsetDateTime
 import java.util.UUID
